+++ lib/b_theme.ml
@@ -22,7 +22,7 @@ DIR = /home/john/.config/bogue/themes
 
 *)
 
-let this_version = "20241114"  (* see VERSION file *)
+let this_version = "20241120"  (* see VERSION file *)
 (* Versions are compared using usual (lexicographic) string ordering. *)
 
 let default_vars = [
@@ -467,6 +467,9 @@ let set_integer_scale b =
 
 (* WARNING: scaling functions should be used after graphics init otherwise
    !scale=0. Use Sync.push  *)
+
+let scale_was_init () = !scale <> 0.
+
 let scale_int i =
   round (!scale *. float i)